feat: add ReportController

+++ sentinel-dashboard/src/main/java/com/alibaba/csp/sentinel/dashboard/controller/v2/report/ReportController.java
@@ -2,7 +2,6 @@ package com.alibaba.csp.sentinel.dashboard.controller.v2.report;
 
 import com.alibaba.csp.sentinel.dashboard.datasource.entity.MetricEntity;
 import com.alibaba.csp.sentinel.dashboard.repository.metric.MetricsRepository;
-import com.alibaba.fastjson.JSON;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,6 +15,7 @@ public class ReportController {
     @Autowired
     private MetricsRepository<MetricEntity> metricStore;
 
+
     @Scheduled(fixedRate = 5000) // 每隔5秒执行一次
     public void test() {
         long endTime = System.currentTimeMillis();
@@ -23,8 +23,8 @@ public class ReportController {
         String app = "graee68d";
         List<MetricEntity> entities = metricStore.queryByAppAndResourceBetween(app, "/pc-service/api/hello", startTime, endTime);
 
-        double sumSuccessQps = entities.stream()
-                .mapToDouble(MetricEntity::getPassQps)
+        long sumSuccessQps = entities.stream()
+                .mapToLong(MetricEntity::getPassQps)
                 .sum();
         System.out.println("sumSuccessQps: " + sumSuccessQps);
     }
